Fix repo list fetching pagination for Github
When user has more than 60 repositories on their user account, the
pagination would silently fail & will repeatedly fetch first & second
pages only. Example of the failure:

response from the header on the second request:
```
<https://api.github.com/user/repos?page=1>; rel="prev",
<https://api.github.com/user/repos?page=3>; rel="next",
<https://api.github.com/user/repos?page=4>; rel="last",
<https://api.github.com/user/repos?page=1>; rel="first"
```

This will parse the next page to be:
```
https://api.github.com/user/repos?page=1>; rel="prev", <https://api.github.com/user/repos?page=3
```

instead of `https://api.github.com/user/repos?page=3`.

The reason it still fetches the first page is that it treats the rest
of the URL as query parameters probably.